Fix Modal.vue background
Currently, when you open a modal e.g. click on the "Delete Account" button from the "Profile" page, the dimmed background extends over the modal content due to a z-order issue.

This is because eslint removes the `transform` class which is no longer needed in Tailwind v3. However, as per the upgrade guide https://tailwindcss.com/docs/upgrade-guide#automatic-transforms-and-filters there is one exception.

This PR replaces `transform` with a `relative` class to fix the issue.